requirements.txt
pandas==2.2.3
huggingface_hub==0.25.1
faiss-cpu==1.9.0
langchain==0.3.3
langchain_community==0.3.2
sentence-transformers==3.2.0